backend/requirements.txt
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
distro==1.7.0
Django==4.0.6
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.3
jsonschema==3.2.0
paramiko==2.11.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
toml==0.10.2
urllib3==1.26.10
websocket-client==0.59.0
